fix: circular dependency because of getUserOs was causing issues in loading files in dev environment
