+++ app/src/utils/Misc.js
@@ -126,29 +126,6 @@ export const parseGravatarImage = (urlString) => {
   return url.href;
 };
 
-export const getUserOS = () => {
-  let userAgent = window.navigator.userAgent,
-    platform = window.navigator.platform,
-    macosPlatforms = ["Macintosh", "MacIntel", "MacPPC", "Mac68K"],
-    windowsPlatforms = ["Win32", "Win64", "Windows", "WinCE"],
-    iosPlatforms = ["iPhone", "iPad", "iPod"],
-    os = null;
-
-  if (macosPlatforms.indexOf(platform) !== -1) {
-    os = "macOS";
-  } else if (iosPlatforms.indexOf(platform) !== -1) {
-    os = "iOS";
-  } else if (windowsPlatforms.indexOf(platform) !== -1) {
-    os = "Windows";
-  } else if (/Android/.test(userAgent)) {
-    os = "Android";
-  } else if (!os && /Linux/.test(platform)) {
-    os = "Linux";
-  }
-
-  return os;
-};
-
 // Modifies the original object! Deep clone prior if required
 export const sanitizeDataForFirebase = (myObject) => {
   // Remove undefined values, if any
